cpp material rework
