+++ game/src/Game/Model/XACLoader.gd
@@ -10,7 +10,7 @@ static var flag_shader : ShaderMaterial = preload("res://src/Game/Model/flag_mat
 
 static var scrolling_shader : ShaderMaterial = preload("res://src/Game/Model/scrolling_mat.tres")
 const MAX_SCROLLING_TEXTURES : int = 32 # max number of textures supported by the shader
-#static var added_scrolling_textures_diffuse : PackedStringArray
+static var added_scrolling_textures_diffuse : PackedStringArray
 const SCROLLING_MATERIAL_FACTORS : Dictionary = {
 	"TexAnim" : 2.5, # Tank tracks
 	"Smoke" : 0.3 # Buildings, factories, steam ships, sieges
@@ -354,8 +354,9 @@ static func make_materials(materialDefinitionChunks : Array[MaterialDefinitionCh
 		if diffuse_name and specular_name:
 			if normal_name:
 				push_error("Normal texture present in unit colours material: ", normal_name)
-
-			var textures_index_spec : int = added_unit_textures_spec.find(specular_name)
+			var textures_index_diffuse = ModelSingleton.set_unit_material_texture(2,diffuse_name)
+			var textures_index_spec = ModelSingleton.set_unit_material_texture(3,specular_name)
+			"""var textures_index_spec : int = added_unit_textures_spec.find(specular_name)
 			if textures_index_spec < 0:
 				var unit_colours_mask_texture : ImageTexture = AssetManager.get_texture(TEXTURES_PATH % specular_name)
 				if unit_colours_mask_texture:
@@ -392,7 +393,8 @@ static func make_materials(materialDefinitionChunks : Array[MaterialDefinitionCh
 					unit_shader.set_shader_parameter(param_texture_diffuse, diffuse_textures)
 				else:
 					push_error("Failed to load diffuse texture: ", diffuse_name)
-
+			"""
+			
 			materials.push_back(MaterialDefinition.new(unit_shader, textures_index_diffuse, textures_index_spec))
 
 		# Flag (diffuse is unionjacksquare which is ignored)
@@ -415,7 +417,7 @@ static func make_materials(materialDefinitionChunks : Array[MaterialDefinitionCh
 			if normal_name:
 				push_error("Normal texture present in scrolling material: ", normal_name)
 			#ModelSingleton. TODO
-			"""var scroll_textures_index_diffuse : int = added_scrolling_textures_diffuse.find(diffuse_name)
+			var scroll_textures_index_diffuse : int = added_scrolling_textures_diffuse.find(diffuse_name)
 			if scroll_textures_index_diffuse < 0:
 				var diffuse_texture : ImageTexture = AssetManager.get_texture(TEXTURES_PATH % diffuse_name)
 				if diffuse_texture:
@@ -438,7 +440,7 @@ static func make_materials(materialDefinitionChunks : Array[MaterialDefinitionCh
 					scroll_factors.push_back(SCROLLING_MATERIAL_FACTORS[matdef.name])
 					scrolling_shader.set_shader_parameter(param_scroll_factor, scroll_factors)
 				else:
-					push_error("Failed to load diffuse texture: ", diffuse_name)"""
+					push_error("Failed to load diffuse texture: ", diffuse_name)
 
 			materials.push_back(MaterialDefinition.new(scrolling_shader, -1, -1, scroll_textures_index_diffuse))
